backend_api/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
scikit-learn==1.3.0
tensorflow==2.13.0
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0
google-search-results==2.4.2
reportlab==4.0.7
psutil==5.9.5
numpy==1.24.0
pandas==2.0.0
python-dotenv==1.0.0
joblib==1.3.2
